+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Unreleased
 
+### Added
+
+- `data-ignored` boolean attribute for detecting elements, that must not be checked for structure difference.
+
 ### Removed
 
 - CommonJS bundle.